feat(actions): promote Actions to a first-class nav surface
Plan gap: the deep-dive page + briefing tributary still required deep-
linking by ID for any action that wasn't currently pending or fired in
the last 24h. No top-level surface meant operators were still hunting.

This adds the missing piece — Actions becomes a persistent nav entry
backed by a tabbed queue covering the full lifecycle:

- SectionNav: "Approvals" → "Actions". Always visible (was conditional
  on pending count > 0). Badge still tracks pending count.
- New /actions index page with tabs: Awaiting you · Fired · Rejected ·
  All. Default = Awaiting you (preserves the current keyboard-driven
  approve/reject flow). Other tabs render a receipt-style card with
  approver, executor result, rejection reason as appropriate. Filter
  state lives in ?filter= so links and reloads are stable.
- /api/approvals: extended with ?filter= (awaiting | fired | rejected
  | all). Response now includes approver kind/label and rejection
  reason so non-pending tabs can render. countOnly behavior unchanged.
- /approvals → permanent redirect to /actions (briefing/finding/old
  links keep working).
- FindingCard's approval cards now route to /actions (was /approvals).

The "first-class" promise from docs/ACTION_RECEIPTS_PLAN.md is now
genuinely true: every action, in every state, is reachable from the
nav in one click.
