+++ apps/web/src/app/api/approvals/route.ts
@@ -1,10 +1,12 @@
 import { NextRequest, NextResponse } from "next/server";
 import {
+  action_policy,
   action_request,
   and,
   db,
   desc,
   eq,
+  inArray,
   observation,
   sql,
   workflow_definition,
@@ -23,11 +25,29 @@ const RISK_ORDER: Record<string, number> = {
   low: 3,
 };
 
+type Filter = "awaiting" | "fired" | "rejected" | "all";
+
+function parseFilter(value: string | null): Filter {
+  if (value === "fired" || value === "rejected" || value === "all") return value;
+  return "awaiting";
+}
+
+function statusesForFilter(filter: Filter): string[] | null {
+  if (filter === "awaiting") return ["pending_approval"];
+  if (filter === "fired") return ["executed", "approved"];
+  if (filter === "rejected") return ["rejected", "failed"];
+  return null; // all
+}
+
 export async function GET(request: NextRequest) {
   const orgId = await getOrgId();
-  const countOnly =
-    new URL(request.url).searchParams.get("countOnly") === "true";
+  const sp = new URL(request.url).searchParams;
+  const countOnly = sp.get("countOnly") === "true";
+  const filter = parseFilter(sp.get("filter"));
 
+  // The nav badge always tracks pending_approval count regardless of the
+  // current view — the operator should always know how many actions need
+  // their call.
   if (countOnly) {
     const [row] = await db()
       .select({ count: sql<number>`count(*)::int` })
@@ -41,9 +61,15 @@ export async function GET(request: NextRequest) {
     return NextResponse.json({ count: row?.count ?? 0 });
   }
 
+  const statuses = statusesForFilter(filter);
+  const statusCondition = statuses
+    ? statuses.length === 1
+      ? eq(action_request.status, statuses[0])
+      : inArray(action_request.status, statuses)
+    : undefined;
+
   const rows = await db()
     .select({
-      // action_request fields
       id: action_request.id,
       workflowRunId: action_request.workflow_run_id,
       triggeredByObservationId: action_request.triggered_by_observation_id,
@@ -53,15 +79,18 @@ export async function GET(request: NextRequest) {
       riskLevel: action_request.risk_level,
       summary: action_request.summary,
       scope: action_request.scope,
+      status: action_request.status,
+      approvedAt: action_request.approved_at,
+      approvedByUserId: action_request.approved_by_user_id,
+      policyId: action_request.policy_id,
+      rejectionReason: action_request.rejection_reason,
       createdAt: action_request.created_at,
-      // workflow_run fields
       runStartedAt: workflow_run.started_at,
       runCreatedAt: workflow_run.created_at,
-      // workflow_definition fields
       workflowId: workflow_definition.id,
       workflowName: workflow_definition.name,
-      // observation fields (nullable)
       observationTitle: observation.title,
+      policyName: action_policy.name,
     })
     .from(action_request)
     .innerJoin(workflow_run, eq(action_request.workflow_run_id, workflow_run.id))
@@ -73,30 +102,35 @@ export async function GET(request: NextRequest) {
       observation,
       eq(action_request.triggered_by_observation_id, observation.id),
     )
-    .where(
-      and(
-        eq(action_request.org_id, orgId),
-        eq(action_request.status, "pending_approval"),
-      ),
+    .leftJoin(
+      action_policy,
+      eq(action_request.policy_id, action_policy.id),
     )
-    .orderBy(desc(action_request.created_at));
+    .where(
+      statusCondition
+        ? and(eq(action_request.org_id, orgId), statusCondition)
+        : eq(action_request.org_id, orgId),
+    )
+    .orderBy(desc(action_request.created_at))
+    .limit(filter === "awaiting" ? 200 : 100);
 
-  // Sort by risk level first (critical → low), then by createdAt desc.
-  const sorted = [...rows].sort((a, b) => {
-    const ra = RISK_ORDER[a.riskLevel ?? ""] ?? 99;
-    const rb = RISK_ORDER[b.riskLevel ?? ""] ?? 99;
-    if (ra !== rb) return ra - rb;
-    return b.createdAt.getTime() - a.createdAt.getTime();
-  });
+  // For the awaiting tab, sort by risk first then time desc — the operator
+  // wants critical at the top. Other tabs are time-ordered (already from SQL).
+  const sorted =
+    filter === "awaiting"
+      ? [...rows].sort((a, b) => {
+          const ra = RISK_ORDER[a.riskLevel ?? ""] ?? 99;
+          const rb = RISK_ORDER[b.riskLevel ?? ""] ?? 99;
+          if (ra !== rb) return ra - rb;
+          return b.createdAt.getTime() - a.createdAt.getTime();
+        })
+      : rows;
 
   return NextResponse.json({
-    approvals: sorted.map((r) => ({
+    actions: sorted.map((r) => ({
       id: r.id,
       workflowRunId: r.workflowRunId,
-      workflow: {
-        id: r.workflowId,
-        name: r.workflowName,
-      },
+      workflow: { id: r.workflowId, name: r.workflowName },
       triggeredByObservation: r.observationTitle
         ? { title: r.observationTitle }
         : null,
@@ -106,9 +140,21 @@ export async function GET(request: NextRequest) {
       riskLevel: r.riskLevel,
       summary: r.summary,
       scope: r.scope,
+      status: r.status,
+      approvedAt: r.approvedAt?.toISOString() ?? null,
+      approverKind: r.approvedByUserId
+        ? ("operator" as const)
+        : r.policyId
+          ? ("policy" as const)
+          : r.approvedAt
+            ? ("auto" as const)
+            : null,
+      approverLabel: r.approvedByUserId ?? r.policyName ?? null,
+      rejectionReason: r.rejectionReason,
       runAt: (r.runStartedAt ?? r.runCreatedAt).toISOString(),
       createdAt: r.createdAt.toISOString(),
     })),
     count: sorted.length,
+    filter,
   });
 }
